refactored structre of the source code, now it's an full feature based structure
